feat: improve fixture to show problem
