+++ projects/tuist/fixtures/app_with_spm_dependencies/App/Sources/App/AppDelegate.swift
@@ -1,0 +1,19 @@
+import AppKit
+import UIKit
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+    var window: UIWindow?
+
+    func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let viewController = UIViewController()
+        viewController.view.backgroundColor = .white
+        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
+
+        AppKit.start()
+
+        return true
+    }
+}
