+++ projects/tuist/fixtures/app_with_spm_dependencies/App/Sources/AppKit/AppKit.swift
@@ -6,19 +6,9 @@ import FirebaseAnalytics
 import FirebaseCore
 import FirebaseDatabase
 import FirebaseFirestore
-import UIKit
 
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-    var window: UIWindow?
-
-    func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
-        window = UIWindow(frame: UIScreen.main.bounds)
-        let viewController = UIViewController()
-        viewController.view.backgroundColor = .white
-        window?.rootViewController = viewController
-        window?.makeKeyAndVisible()
-
+public struct AppKit {
+    public static func start() {
         // Use Alamofire to make sure it links fine
         _ = AF.download("http://www.tuist.io")
 
@@ -36,11 +26,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
         // Use FirebaseFirestore to make sure it links fine
         _ = Firestore.firestore()
-
-        return true
-    }
-
-    func hello() -> String {
-        "AppDelegate.hello()"
     }
 }
